+++ src/memory/pool/BasicMemoryPool.h
@@ -15,14 +15,14 @@ namespace pipey {
 				virtual ~CBasicMemoryPool(void) { };
 
 			public:
-				virtual void Allocate(unsigned long nSize, CMemoryHandle &rHandle);
+				virtual void Allocate(uint32_t nSize, CMemoryHandle &rHandle);
 				virtual void CloseHandle(CMemoryHandle & rHandle);
 				virtual void DuplicateHandle(const CMemoryHandle & rSource, CMemoryHandle & rTarget);
 
 				virtual char *GetBuffer(const CMemoryHandle &rHandle);
 
 			private:
-				std::map<char *, unsigned long> m_handles;
+				std::map<char *, uint32_t> m_handles;
 			};
 		}
 	}
