+++ src/thread/sync/TimerableLockPtr.h
@@ -23,7 +23,7 @@ namespace pipey {
 
 
 			public:
-				virtual pipey::common::SYNC_RESULT AcquireTimedLock(unsigned long nMilliSeconds = pipey::common::TIME_INFINITE);
+				virtual pipey::common::SYNC_RESULT AcquireTimedLock(uint32_t nMilliSeconds = pipey::common::TIME_INFINITE);
 
 			};
